requirements.txt
absl-py==0.11.0
alembic==1.6.5
aniso8601==8.0.0
appnope==0.1.0
asn1crypto==0.24.0
attrs==21.2.0
autocommand==2.2.1
Automat==20.2.0
backcall==0.1.0
beautifulsoup4==4.9.1
biopython==1.77
bleach==3.1.0
blinker==1.4
bokeh==1.3.4
brewer2mpl==1.4.1
cached-property==1.5.2
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
charset-normalizer==2.0.4
chart-studio==1.0.0
Click==7.0
click-conf==0.0.1
click-stream==0.0.10
colorlover==0.3.0
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==2.7
cssselect==1.1.0
cufflinks==0.16
cycler==0.10.0
Cython==0.29.21
dash==1.4.1
dash-core-components==1.3.1
dash-daq==0.2.1
dash-html-components==1.0.1
dash-renderer==1.1.2
dash-table==4.4.1
dash.ly==0.17.3
debugpy==1.3.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
elastic2-doc-manager==1.0.0
elasticsearch==7.14.0
elasticsearch-loader==0.6.0
entrypoints==0.3
fake-useragent==0.1.11
filelock==3.0.12
Flask==1.1.1
Flask-Compress==1.4.0
Flask-JWT-Extended==4.2.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
flask-restx==0.2.0
Flask-SeaSurf==0.2.2
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.18.1
ggplot==0.11.5
h5py==3.3.0
huggingface-hub==0.0.12
hyperlink==19.0.0
idna==2.8
importlib-metadata==3.10.1
importlib-resources==5.2.2
incremental==17.5.0
iniconfig==1.1.1
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.10.1
joblib==0.14.1
json262==0.2.0
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.0.0
jupyter-core==4.6.3
kiwisolver==1.1.0
lvis==0.5.3
lxml==4.5.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.1.1
matplotlib-inline==0.1.2
mistune==0.8.4
mongo-connector==3.1.1
mysql-connector-python==8.0.16
mysql-connector-python-rf==2.2.2
mysqlclient==1.4.2.post1
nbconvert==5.6.0
nbformat==4.4.0
nltk==3.6.2
notebook==6.0.0
numpy==1.18.5
object-detection==0.1
opencv-python==4.5.1.48
packaging==21.0
pandas==1.0.5
pandocfilters==1.4.2
parsel==1.6.0
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickle5==0.0.10
pickleshare==0.7.5
Pillow==6.2.1
plotly==4.1.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Protego==0.1.16
protobuf==3.8.0
ptyprocess==0.6.0
py==1.10.0
py4j==0.10.7
pyaml==21.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==2.1.0
pykerberos==1.2.1
pymongo==3.11.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.2
pyrsistent==0.15.4
pyspark==2.4.3
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-json-logger==2.0.2
python-snappy==0.6.0
pytz==2019.1
PyYAML==5.1.2
pyzmq==19.0.1
qtconsole==4.5.3
queuelib==1.5.0
regex==2020.2.20
requests==2.26.0
retrying==1.3.3
sacremoses==0.0.45
scikit-learn==0.22.2.post1
scipy==1.3.0
Scrapy==2.1.0
Scrapy-UserAgents==0.0.1
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
soupsieve==2.0.1
SQLAlchemy==1.3.5
statsmodels==0.10.1
terminado==0.8.3
testpath==0.4.2
tf-slim==1.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.0
tornado==6.0.4
tqdm==4.61.2
traitlets==4.3.3
transformers==4.8.2
Twisted==20.3.0
typing-extensions==3.10.0.0
urllib3==1.25.11
virtualenv==16.6.1
w3lib==1.22.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.1
WTForms==2.2.1
xmltodict==0.12.0
zipp==3.5.0
zope.interface==5.1.0